chore: create shared table function
